change reduce file

+++ reduce.py
@@ -2,4 +2,4 @@ from functools import reduce
 
 def factors(n):    
     return set(reduce(list.__add__, 
-                ([i, n//i] for i in range(1, int(pow(n, 0.5) + 1)) if n % i == 0)))
+                ([i, n//i] for i in range(1, long(pow(n, 0.5) + 1)) if n % i == 0)))
